+++ hypervisor/debug/sbuf.c
@@ -153,9 +153,9 @@ int sbuf_put(struct shared_buf *sbuf, uint8_t *data)
 	return sbuf->ele_size;
 }
 
-int sbuf_share_setup(uint32_t pcpu_id, uint32_t sbuf_id, uint64_t *hva)
+int sbuf_share_setup(uint16_t pcpu_id, uint32_t sbuf_id, uint64_t *hva)
 {
-	if (pcpu_id >= (uint32_t) phy_cpu_num ||
+	if (pcpu_id >= phy_cpu_num ||
 			sbuf_id >= ACRN_SBUF_ID_MAX)
 		return -EINVAL;
