+++ hypervisor/include/arch/x86/lapic.h
@@ -158,14 +158,14 @@ struct lapic_regs {
 void write_lapic_reg32(uint32_t offset, uint32_t value);
 void save_lapic(struct lapic_regs *regs);
 int early_init_lapic(void);
-int init_lapic(uint32_t cpu_id);
+int init_lapic(uint16_t cpu_id);
 void send_lapic_eoi(void);
 uint32_t get_cur_lapic_id(void);
 int send_startup_ipi(enum intr_cpu_startup_shorthand cpu_startup_shorthand,
 		uint32_t cpu_startup_dest,
 		uint64_t cpu_startup_start_address);
 /* API to send an IPI to a single guest */
-void send_single_ipi(uint32_t pcpu_id, uint32_t vector);
+void send_single_ipi(uint16_t pcpu_id, uint32_t vector);
 
 void suspend_lapic(void);
 void resume_lapic(void);
